리팩토링 - 게시판 삭제 완료
